truvade-backend/pyproject.toml
[project]
name = "truvade-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "boto3>=1.42.83",
    "django>=6.0.3",
    "django-storages>=1.14.6",
    "djangorestframework>=3.17.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "gunicorn>=25.3.0",
    "httpx>=0.28.1",
    "model-bakery>=1.23.4",
    "pillow>=12.2.0",
    "psycopg2-binary>=2.9.11",
    "pytest>=9.0.2",
    "pytest-django>=4.12.0",
]
